services/admin_service/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
alembic
psycopg2-binary
dotenv
httpx
pydantic[email]
python-slugify
python-jose
redis[hiredis]>=4.2.0